+++ tests/test_metrics.py
@@ -1,15 +1,12 @@
 from pathlib import Path
 
-import matplotlib
+import numpy
 import pandas
+import pytest
+from matplotlib import pyplot
 from numpy.random.mtrand import RandomState
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.tree import DecisionTreeClassifier
-
-matplotlib.use('agg')
-import numpy
-import pytest
-from matplotlib import pyplot
 
 from ds_utils.metrics import plot_confusion_matrix, plot_metric_growth_per_labeled_instances, \
     visualize_accuracy_grouped_by_probability